Upload web scraper for food recalls on USDA
Scrapes the title, status, date, reason, impacted products, description, and article link of the food recalls posted on the USDA recall website: https://www.fsis.usda.gov/recalls and puts into a data frame. The code has the option to be presented in a loop by uncommenting the printf statement at the end of the while loop.

The end of the notebook contains a simple value count of the different reasons for recalls.